+++ erpnext/accounts/doctype/payment_request/payment_request.json
@@ -461,7 +461,6 @@
    "read_only": 1
   },
   {
-   "default": "0",
    "fieldname": "payment_entry_creation_error",
    "fieldtype": "Small Text",
    "label": "Payment Entry Creation Error",
@@ -472,7 +471,7 @@
  "grid_page_length": 500,
  "is_submittable": 1,
  "links": [],
- "modified": "2025-09-10 14:43:23.947398",
+ "modified": "2025-09-10 15:16:37.242216",
  "modified_by": "Administrator",
  "module": "Accounts",
  "name": "Payment Request",
