+++ erpnext/accounts/doctype/sales_invoice/sales_invoice.py
@@ -792,25 +792,6 @@ class SalesInvoice(SellingController):
 	def get_company_abbr(self):
 		return frappe.db.sql("select abbr from tabCompany where name=%s", self.company)[0][0]
 
-	def validate_debit_to_acc(self):
-		account = frappe.get_cached_value("Account", self.debit_to,
-			["account_type", "report_type", "account_currency"], as_dict=True)
-
-		if not account:
-			frappe.throw(_("Debit To is required"), title=_("Account Missing"))
-
-		if account.report_type != "Balance Sheet":
-			frappe.throw(_("Please ensure {} account is a Balance Sheet account. \
-					You can change the parent account to a Balance Sheet account or select a different account.")
-				.format(frappe.bold("Debit To")), title=_("Invalid Account"))
-
-		if self.customer and account.account_type != "Receivable":
-			frappe.throw(_("Please ensure {} account is a Receivable account. \
-					Change the account type to Receivable or select a different account.")
-				.format(frappe.bold("Debit To")), title=_("Invalid Account"))
-
-		self.party_account_currency = account.account_currency
-
 	def validate_with_previous_doc(self):
 		sales_order_compare = [["currency", "="], ["branch", "="]]
 		delivery_note_compare = [["company", "="], ["branch", "="], ["currency", "="]]
